feat(compiler): type classes, given instances, UFCS dispatch, operator overloading, and UDN
Type class system:
- Add Class/Given/Satisfies/Where/Law keywords (lexer)
- Add ClassDef/GivenDef/Constraint/ClassMember AST nodes
- Parse class/given/where/law/satisfies; parser now 47/47 tests
- New opcode: CallMethod {method_idx, arg_count} (0x83)
- Module.method_table: runtime method registry by name+type_tag
- UFCS routing: record field > free fn > class method
- VM exec_call_method: dispatch by receiver type_tag
- Intrinsics: Writeln=0, Write=1, IntToString=2, FloatToString=3
- Std: Eq/Ord/Add/Sub/Mul/Div/Rem/Show classes + given for Int/Float/String
- Std: given Eq[Bool], given Div[Float]
- Examples 01-10: subdirs, all verified working

Operator dispatch for user-defined types:
- Value::Object gains type_tag: u16 (0=anon, 10+=user-defined)
- NewObj opcode: 5-byte encoding with type_tag + field_count
- Emitter assigns type tags sequentially starting at 10
- VM binary op handlers (AddI64, SubI64, etc.) fall back to method_table
  when LHS is a user-defined Object type — no new opcodes needed
- track_type_of_binding: propagates type through UFCS calls (v3 := v1.add(v2))
- Example 10: Vec2 + given Add[Vec2] — v1.add(v2) dispatches correctly

Universal Dot Notation (UDN):
- Lexer: Question ?, QuestionDot ?., QuestionQuestion ?? tokens
- AST: Expr::DotPrefix, PostfixOp::OptField, BinOp::NilCoalesce, Ty::Option
- Parser: ?T sugar, .Name(args) dot-prefix, ?.field optional chain, ?? nil coal
- Codegen: NilCoalesce opcode (0x84), OPT_FIELD placeholder (0x85)
- VM: exec_nil_coalesce, exec_opt_field handlers
- Grammar: ast_ty_option, ast_expr_dot_prefix, ast_opt_field, ast_expr_nil_coal
- VS Code: ?. and ?? operator highlighting

+++ crates/musi_codegen/src/module.rs
@@ -164,8 +164,8 @@ impl FunctionEntry {
 pub struct MethodEntry {
     /// Method name, e.g. `"eq"`.
     pub name: Box<str>,
-    /// Type tag of the receiver (see `TypeTag` in the emitter).
-    pub type_tag: u8,
+    /// Type tag of the receiver (0–9 = primitives, 10+ = user-defined types).
+    pub type_tag: u16,
     /// Index into `function_table`.
     pub fn_idx: u16,
 }
